components: 🔄 Refactored for better file structure
There were becoming way too many component files. This should make things a lot cleaner
