+++ app/components/EditEvents/EditEvents.scss
