+++ app/components/ReaderMapForm/ReaderFields.js
@@ -1,10 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import DropdownSelect from './DropdownSelect';
-import { readerShape } from './SyncReaders';
+import DropdownSelect from '../DropdownSelect';
 
-const { shape, string, bool, func, arrayOf } = PropTypes;
+const { string, bool, func, arrayOf } = PropTypes;
 
 const ReaderFields = (props) => (
   <div>
@@ -56,32 +55,4 @@ ReaderFields.propTypes = {
   onRemoveReader: func.isRequired,
 };
 
-const ReaderMapForm = (props) => {
-  const { readerMap, ...readerFieldProps } = props;
-  return (
-    <div>
-      {readerMap.length > 0 &&
-        readerMap.map(({ address, event }) => (
-          <ReaderFields
-            key={`reader-${address}`}
-            {...{ address, event, ...readerFieldProps }}
-          />
-        ))
-      }
-    </div>
-  );
-};
-
-ReaderMapForm.propTypes = {
-  readerMap: arrayOf(shape(readerShape)).isRequired,
-  eventList: arrayOf(string),
-  onChangeValue: func.isRequired,
-  onRemoveReader: func.isRequired
-};
-
-ReaderMapForm.defaultProps = {
-  readerMap: {},
-  eventList: []
-};
-
-export default ReaderMapForm;
+export default ReaderFields;
