+++ app/components/Timer/Timer.scss
