+++ app/components/Timer/index.js
